add 'missing' casts

+++ code/network/multi.cpp
@@ -1499,7 +1499,7 @@ void multi_do_frame()
 		
 		// ping everyone
 		multi_ping_send_all();
-		Next_ping_time = time(NULL);		
+		Next_ping_time = (int) time(NULL);		
 	}	
 	
 	// if I am the master, and we are not yet actually playing the mission, send off netgame
@@ -1510,7 +1510,7 @@ void multi_do_frame()
 			if ( (Netgame_send_time < 0) || ((time(NULL) - Netgame_send_time) > NETGAME_SEND_TIME) ) {
 				send_netgame_update_packet();				
 				
-				Netgame_send_time = time(NULL);
+				Netgame_send_time = (int) time(NULL);
 			}		
 		} else {
 			if ( (State_send_time < 0) || ((time(NULL) - State_send_time) > STATE_SEND_TIME) ){
@@ -1519,7 +1519,7 @@ void multi_do_frame()
 					send_netplayer_update_packet();
 				}				
 				
-				State_send_time = time(NULL);
+				State_send_time = (int) time(NULL);
 			}
 		}
 	}
@@ -1529,7 +1529,7 @@ void multi_do_frame()
 			if ( (Gameinfo_send_time < 0) || ((time(NULL) - Gameinfo_send_time) > GAMEINFO_SEND_TIME)){
 				send_game_info_packet();
 				
-				Gameinfo_send_time = time(NULL);
+				Gameinfo_send_time = (int) time(NULL);
 			}
 			
 			// for any potential respawns
@@ -1585,7 +1585,7 @@ void multi_do_frame()
 				}
 
 				// reset timestamp
-				Next_bytes_time = time(NULL);				
+				Next_bytes_time = (int) time(NULL);				
 			}
 		} else {			
 			// sending new objects from here is dependent on having objects only created after
@@ -1683,7 +1683,7 @@ void multi_pause_do_frame()
 		if ( (Gameinfo_send_time < 0) || ((time(NULL) - Gameinfo_send_time) > GAMEINFO_SEND_TIME) ){
 			send_game_info_packet();
 			
-			Gameinfo_send_time = time(NULL);
+			Gameinfo_send_time = (int) time(NULL);
 		}				
 	}
 
@@ -1691,7 +1691,7 @@ void multi_pause_do_frame()
 	if((Next_ping_time < 0) || ((time(NULL) - Next_ping_time) > PING_SEND_TIME) ){
 		multi_ping_send_all();
 		
-		Next_ping_time = time(NULL);
+		Next_ping_time = (int) time(NULL);
 	}
 
 	// periodically send a client update packet to all clients
